ast_results/github_android/app/src/main/java/com/github/pockethub/android/ui/MainActivity.java
// isComment
package com.github.pockethub.android.ui;

import android.accounts.Account;
import android.accounts.AccountManager;
import android.app.SearchManager;
import android.content.Context;
import android.content.Intent;
import android.content.SharedPreferences;
import android.content.res.Configuration;
import android.os.Build;
import android.os.Bundle;
import android.preference.PreferenceManager;
import android.support.annotation.NonNull;
import android.support.annotation.VisibleForTesting;
import android.support.design.widget.NavigationView;
import android.support.v4.app.Fragment;
import android.support.v4.app.FragmentManager;
import android.support.v4.view.GravityCompat;
import android.support.v4.view.MenuItemCompat;
import android.support.v4.widget.DrawerLayout;
import android.support.v7.app.ActionBarDrawerToggle;
import android.support.v7.widget.SearchView;
import android.util.Log;
import android.view.Menu;
import android.view.MenuItem;
import android.view.SubMenu;
import android.view.View;
import android.view.Window;
import android.webkit.CookieManager;
import android.widget.ImageView;
import android.widget.TextView;
import com.github.pockethub.android.R;
import com.github.pockethub.android.accounts.AccountUtils;
import com.github.pockethub.android.accounts.AccountsHelper;
import com.github.pockethub.android.accounts.LoginActivity;
import com.github.pockethub.android.core.user.UserComparator;
import com.github.pockethub.android.persistence.AccountDataManager;
import com.github.pockethub.android.persistence.CacheHelper;
import com.github.pockethub.android.rx.AutoDisposeUtils;
import com.github.pockethub.android.ui.gist.GistsPagerFragment;
import com.github.pockethub.android.ui.issue.FilterListFragment;
import com.github.pockethub.android.ui.issue.IssueDashboardPagerFragment;
import com.github.pockethub.android.ui.notification.NotificationActivity;
import com.github.pockethub.android.ui.user.HomePagerFragment;
import com.github.pockethub.android.util.AvatarLoader;
import com.github.pockethub.android.util.ToastUtils;
import com.meisolsson.githubsdk.core.TokenStore;
import com.meisolsson.githubsdk.model.User;
import java.util.Collections;
import java.util.HashMap;
import java.util.List;
import java.util.Map;
import javax.inject.Inject;
import javax.inject.Provider;
import javax.inject.Singleton;
import butterknife.BindView;
import io.reactivex.Single;
import io.reactivex.android.schedulers.AndroidSchedulers;
import io.reactivex.schedulers.Schedulers;

public class isClassOrIsInterface extends BaseActivity implements NavigationView.OnNavigationItemSelectedListener {

    private static final String isVariable = "isStringConstant";

    private static final String isVariable = "isStringConstant";

    @BindView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    protected DrawerLayout isVariable;

    @BindView(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)
    protected NavigationView isVariable;

    @Inject
    protected AccountDataManager isVariable;

    @Inject
    protected Provider<UserComparator> isVariable;

    private List<User> isVariable = isNameExpr.isMethod();

    private User isVariable;

    @Inject
    @Singleton
    protected AvatarLoader isVariable;

    private boolean isVariable;

    private ActionBarDrawerToggle isVariable;

    @VisibleForTesting
    public Fragment isVariable;

    @Override
    protected void isMethod(Bundle isParameter) {
        super.isMethod(isNameExpr);
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        SharedPreferences isVariable = isNameExpr.isMethod(this);
        isNameExpr = isNameExpr.isMethod(isNameExpr, true);
        isMethod().isMethod(true);
        isNameExpr = new ActionBarDrawerToggle(this, isNameExpr, isMethod(), isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {

            @Override
            public void isMethod(View isParameter) {
                super.isMethod(isNameExpr);
                if (!isNameExpr) {
                    SharedPreferences isVariable = isNameExpr.isMethod(isNameExpr.this);
                    isNameExpr.isMethod().isMethod(isNameExpr, true).isMethod();
                    isNameExpr = true;
                    isNameExpr.isMethod(isNameExpr, "isStringConstant");
                }
            }
        };
        isNameExpr.isMethod(isNameExpr);
        isNameExpr.isMethod(this);
        isMethod();
        TokenStore isVariable = isNameExpr.isMethod(this);
        if (isNameExpr.isMethod() == null) {
            AccountManager isVariable = isNameExpr.isMethod(this);
            Account[] isVariable = isNameExpr.isMethod(isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
            if (isNameExpr.isFieldAccessExpr > isIntegerConstant) {
                Account isVariable = isNameExpr[isIntegerConstant];
                isNameExpr.isMethod(this, isNameExpr);
                isNameExpr.isMethod(isNameExpr.isMethod(this, isNameExpr));
            }
        }
    }

    @Override
    public void isMethod(Configuration isParameter) {
        super.isMethod(isNameExpr);
        isNameExpr.isMethod(isNameExpr);
    }

    @Override
    public boolean isMethod(MenuItem isParameter) {
        if (isNameExpr.isMethod(isNameExpr)) {
            return true;
        }
        return super.isMethod(isNameExpr);
    }

    @Override
    protected void isMethod(Bundle isParameter) {
        super.isMethod(isNameExpr);
        isNameExpr.isMethod();
    }

    private void isMethod() {
        isNameExpr.isMethod(() -> isNameExpr.isMethod(isMethod(), this)).isMethod(isParameter -> isNameExpr.isMethod(true)).isMethod(isParameter -> {
            isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod());
            return isNameExpr;
        }).isMethod(isNameExpr.isMethod()).isMethod(isNameExpr.isMethod()).isMethod(isNameExpr.isMethod(this)).isMethod(this::onOrgsLoaded, isParameter -> {
            isNameExpr.isMethod(isNameExpr, "isStringConstant", isNameExpr);
            isNameExpr.isMethod(this, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        });
    }

    @VisibleForTesting
    void isMethod(List<User> isParameter) {
        if (isNameExpr.isMethod()) {
            return;
        }
        isNameExpr = isNameExpr.isMethod(isIntegerConstant);
        this.isFieldAccessExpr = isNameExpr;
        isMethod();
        Window isVariable = isMethod();
        if (isNameExpr == null) {
            return;
        }
        View isVariable = isNameExpr.isMethod();
        if (isNameExpr == null) {
            return;
        }
        isNameExpr.isMethod(() -> {
            isMethod(new HomePagerFragment(), isNameExpr);
            if (!isNameExpr) {
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
            }
        });
    }

    @Override
    public boolean isMethod(Menu isParameter) {
        isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr);
        SearchManager isVariable = (SearchManager) isMethod(isNameExpr.isFieldAccessExpr);
        MenuItem isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        SearchView isVariable = (SearchView) isNameExpr.isMethod(isNameExpr);
        isNameExpr.isMethod(isNameExpr.isMethod(isMethod()));
        return super.isMethod(isNameExpr);
    }

    @Override
    protected void isMethod() {
        super.isMethod();
        // isComment
        // isComment
        List<User> isVariable = isNameExpr;
        if (isNameExpr != null && !isNameExpr.isMethod() && !isNameExpr.isMethod(this, isNameExpr.isMethod(isIntegerConstant))) {
            isMethod();
        }
    }

    Map<MenuItem, User> isVariable = new HashMap<>();

    private void isMethod() {
        ImageView isVariable;
        TextView isVariable;
        TextView isVariable;
        View isVariable = isNameExpr.isMethod(isIntegerConstant);
        isNameExpr = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        ImageView isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr.isMethod(isParameter -> isMethod(new Intent(isNameExpr.this, NotificationActivity.class)));
        isNameExpr.isMethod(isNameExpr, isNameExpr);
        isNameExpr.isMethod(isNameExpr.isMethod());
        String isVariable = isNameExpr.isMethod();
        if (isNameExpr != null) {
            isNameExpr.isMethod(isNameExpr.isMethod());
        } else {
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
        }
    }

    private void isMethod() {
        isMethod();
        isMethod();
    }

    private void isMethod() {
        MenuItem isVariable = isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr.isMethod()) {
            SubMenu isVariable = isNameExpr.isMethod();
            for (int isVariable = isIntegerConstant; isNameExpr < isNameExpr.isMethod(); isNameExpr++) {
                User isVariable = isNameExpr.isMethod(isNameExpr);
                if (isNameExpr.isMethod(isNameExpr.isMethod().isMethod()) == null) {
                    MenuItem isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod().isMethod(), isNameExpr.isFieldAccessExpr, isNameExpr.isMethod() != null ? isNameExpr.isMethod() : isNameExpr.isMethod());
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    // isComment
                    // isComment
                    isNameExpr.isMethod(isNameExpr, isNameExpr);
                }
            }
        } else {
            throw new IllegalStateException("isStringConstant" + isNameExpr + "isStringConstant");
        }
    }

    @Override
    public boolean isMethod(@NonNull MenuItem isParameter) {
        int isVariable = isNameExpr.isMethod();
        if (isNameExpr == isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
            isMethod(new HomePagerFragment(), isNameExpr);
            isMethod().isMethod(isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
            return true;
        } else if (isNameExpr == isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
            isMethod(new GistsPagerFragment(), null);
            isMethod().isMethod(isNameExpr.isMethod());
            return true;
        } else if (isNameExpr == isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
            isMethod(new IssueDashboardPagerFragment(), null);
            isMethod().isMethod(isNameExpr.isMethod());
            return true;
        } else if (isNameExpr == isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
            isMethod(new FilterListFragment(), null);
            isMethod().isMethod(isNameExpr.isMethod());
            return true;
        } else if (isNameExpr == isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
            isMethod();
            return true;
        } else if (isNameExpr.isMethod(isNameExpr)) {
            isMethod(new HomePagerFragment(), isNameExpr.isMethod(isNameExpr));
            isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod(true);
            return true;
        } else {
            throw new IllegalStateException("isStringConstant" + isNameExpr + "isStringConstant");
        }
    }

    private void isMethod() {
        // isComment
        if (isNameExpr.isFieldAccessExpr.isFieldAccessExpr >= isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
            isNameExpr.isMethod().isMethod(null);
        } else {
            isNameExpr.isMethod().isMethod();
        }
        // isComment
        CacheHelper isVariable = new CacheHelper(this);
        isNameExpr.isMethod().isMethod("isStringConstant", null, null);
        isNameExpr.isMethod().isMethod("isStringConstant", null, null);
        isNameExpr.isMethod().isMethod("isStringConstant", null, null);
        // isComment
        AccountManager isVariable = isMethod();
        String isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        Account[] isVariable = isNameExpr.isMethod(isNameExpr);
        for (Account isVariable : isNameExpr) {
            if (isNameExpr.isFieldAccessExpr.isFieldAccessExpr >= isNameExpr.isFieldAccessExpr.isFieldAccessExpr) {
                isNameExpr.isMethod(isNameExpr, this, isParameter -> isMethod(), null);
            } else {
                isNameExpr.isMethod(isNameExpr, isParameter -> isMethod(), null);
            }
        }
    }

    private void isMethod() {
        Intent isVariable = new Intent(this, LoginActivity.class);
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr | isNameExpr.isFieldAccessExpr);
        isMethod(isNameExpr);
        isMethod();
    }

    private void isMethod(Fragment isParameter, User isParameter) {
        if (isNameExpr != null) {
            Bundle isVariable = new Bundle();
            isNameExpr.isMethod("isStringConstant", isNameExpr);
            isNameExpr.isMethod(isNameExpr);
        }
        FragmentManager isVariable = isMethod();
        isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr).isMethod();
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
        isNameExpr = isNameExpr;
    }

    private AccountManager isMethod() {
        return isNameExpr.isMethod(this);
    }
}
